improved medication duplicate filtering
